cloneable, bifurcated pubsub, moved some files
