+++ 0e.algo-cyan/src/lib/infra/Cloneable.ts
@@ -1,0 +1,3 @@
+export interface Cloneable<T> {
+  clone(): T;
+}
