+++ 0e.algo-cyan/src/lib/infra/PubSub.ts
@@ -1,5 +1,5 @@
 export class PubSub<T> {
-  private subscribers: Set<(data: T) => void>;
+  protected subscribers: Set<(data: T) => void>;
 
   constructor() {
     this.subscribers = new Set();
